+++ modules/alloy/src/main/ts/ephox/alloy/positioning/layout/LayoutTypes.ts
@@ -1,5 +1,5 @@
-import { Bubble } from './Bubble';
 import { SpotInfo } from '../view/SpotInfo';
+import { Bubble } from './Bubble';
 
 export interface AnchorBox {
   x: () => number;