+++ modules/alloy/src/main/ts/ephox/alloy/positioning/view/Bounder.ts
@@ -1,7 +1,8 @@
 import { Adt, Arr, Fun, Num } from '@ephox/katamari';
-import { Bounds } from '../../alien/Boxes';
+import * as Boxes from '../../alien/Boxes';
 import { Bubble } from '../layout/Bubble';
 import * as Direction from '../layout/Direction';
+import * as LayoutBounds from '../layout/LayoutBounds';
 import { AnchorBox, AnchorElement, AnchorLayout } from '../layout/LayoutTypes';
 import * as Reposition from './Reposition';
 import { SpotInfo } from './SpotInfo';
@@ -26,7 +27,7 @@ const adt: {
   { nofit: [ 'reposition', 'deltaW', 'deltaH' ] }
 ]);
 
-const calcReposition = (newX: number, newY: number, width: number, height: number, bounds: Bounds) => {
+const calcReposition = (newX: number, newY: number, width: number, height: number, bounds: Boxes.Bounds) => {
   const boundsX = bounds.x();
   const boundsY = bounds.y();
   const boundsWidth = bounds.width();
@@ -46,12 +47,15 @@ const calcReposition = (newX: number, newY: number, width: number, height: numbe
   const deltaW = Math.abs(Math.min(width, xInBounds ? boundsX + boundsWidth - newX : boundsX - (newX + width)));
   const deltaH = Math.abs(Math.min(height, yInBounds ? boundsY + boundsHeight - newY : boundsY - (newY + height)));
 
-  // TBIO-3366 + TBIO-4236:
-  // Futz with the X position to ensure that x is positive, but not off the right side of the screen.
+  // measure the maximum x and y, taking into account the height and width of the element
+  const maxX = Math.max(bounds.x(), bounds.right() - width);
+  const maxY = Math.max(bounds.y(), bounds.bottom() - height);
+
+  // Futz with the X value to ensure that we're not off the left or right of the screen
   // NOTE: bounds.x() is 0 in repartee here.
-  const limitX = Num.clamp(newX, bounds.x(), bounds.right());
-  // Futz with the Y value to ensure that we're not off the top of the screen
-  const limitY = Num.clamp(newY, bounds.y(), bounds.bottom());
+  const limitX = Num.clamp(newX, bounds.x(), maxX);
+  // Futz with the Y value to ensure that we're not off the top or bottom of the screen
+  const limitY = Num.clamp(newY, bounds.y(), maxY);
 
   return {
     originInBounds,
@@ -63,23 +67,25 @@ const calcReposition = (newX: number, newY: number, width: number, height: numbe
   };
 };
 
-const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bounds): BounderAttemptAdt  => {
+const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Boxes.Bounds): BounderAttemptAdt  => {
   const candidateX = candidate.x();
   const candidateY = candidate.y();
-  const bubbleLeft = candidate.bubble().offset().left();
-  const bubbleTop = candidate.bubble().offset().top();
+  const bubbleOffsets = candidate.bubble().offset();
+  const bubbleLeft = bubbleOffsets.left();
+  const bubbleTop = bubbleOffsets.top();
 
-  const boundsY = bounds.y();
-  const boundsBottom = bounds.bottom();
-
-  const boundsX = bounds.x();
-  const boundsRight = bounds.right();
+  // adjust the bounds to account for the layout and bubble restrictions
+  const adjustedBounds = LayoutBounds.adjustBounds(bounds, candidate.boundsRestriction(), bubbleOffsets);
+  const boundsY = adjustedBounds.y();
+  const boundsBottom = adjustedBounds.bottom();
+  const boundsX = adjustedBounds.x();
+  const boundsRight = adjustedBounds.right();
 
   // candidate position is excluding the bubble, so add those values as well
   const newX = candidateX + bubbleLeft;
   const newY = candidateY + bubbleTop;
 
-  const { originInBounds, sizeInBounds, limitX, limitY, deltaW, deltaH } = calcReposition(newX, newY, width, height, bounds);
+  const { originInBounds, sizeInBounds, limitX, limitY, deltaW, deltaH } = calcReposition(newX, newY, width, height, adjustedBounds);
 
   // TBIO-3367 + TBIO-3387:
   // Futz with the "height" of the popup to ensure if it doesn't fit it's capped at the available height.
@@ -113,17 +119,18 @@ const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bou
   // console.log('xfit', (boundsX + boundsWidth), ',', (newX + width), ',', newX);
   // console.log('yfit', (boundsY + boundsHeight), ',', (newY + height), ',', newY, ',', height);
   // console.table([{
-  //   xInBounds,
-  //   xFit,
-  //   limitX,
-  //   minX,
-  //   maxX,
-  //   deltaW,
-  //   boundsX,
-  //   boundsWidth,
-  //   candidateX: candidate.x(),
+  //   newY,
+  //   limitY,
+  //   boundsY,
+  //   boundsBottom,
   //   newX,
-  //   width
+  //   limitX,
+  //   boundsX,
+  //   boundsRight,
+  //   candidateX: candidate.x(),
+  //   candidateY: candidate.y(),
+  //   width,
+  //   height
   // }]);
   // console.log('y', yInBounds, yFit, '\t', Math.round(deltaH), '\t', (boundsY === 0 ? '000' : Math.round(boundsY)), '\t', Math.round(boundsHeight), '\t', Math.round(candidate.y()), '\t', Math.round(newY), '\t', height);
   // console.log('maxheight:', deltaH, maxHeight);
@@ -145,7 +152,7 @@ const attempt = (candidate: SpotInfo, width: number, height: number, bounds: Bou
  * bubbles: the bubbles for the popup (see api.Bubble)
  * bounds: the screen
  */
-const attempts = (candidates: AnchorLayout[], anchorBox: AnchorBox, elementBox: AnchorElement, bubbles: Bubble, bounds: Bounds): Reposition.RepositionDecision => {
+const attempts = (candidates: AnchorLayout[], anchorBox: AnchorBox, elementBox: AnchorElement, bubbles: Bubble, bounds: Boxes.Bounds): Reposition.RepositionDecision => {
   const panelWidth = elementBox.width();
   const panelHeight = elementBox.height();
   const attemptBestFit = (layout: AnchorLayout, reposition: Reposition.RepositionDecision, deltaW: number, deltaH: number) => {