+++ modules/alloy/src/test/ts/atomic/position/view/BounderCalcRepositionTest.ts
@@ -6,18 +6,19 @@ import * as Bounder from 'ephox/alloy/positioning/view/Bounder';
 
 UnitTest.test('BounderCalcRepositionTest', () => {
 
-  const nonZeroArb = Jsc.integer(10, 1000);
-  const zeroableArb = Jsc.integer(0, 1000);
+  const maxBounds = 2000;
+  const minBounds  = 0;
+  const zeroableArb = Jsc.integer(minBounds, maxBounds);
 
   const arbTestCase = Jsc.bless({
-    generator: zeroableArb.generator.flatMap((newX: number) => {
-      return zeroableArb.generator.flatMap((newY: number) => {
-        return nonZeroArb.generator.flatMap((width: number) => {
-          return nonZeroArb.generator.flatMap((height: number) => {
-            return zeroableArb.generator.flatMap((boundsX: number) => {
-              return zeroableArb.generator.flatMap((boundsY: number) => {
-                return zeroableArb.generator.flatMap((boundsW: number) => {
-                  return zeroableArb.generator.map((boundsH: number) => {
+    generator: zeroableArb.generator.flatMap((boundsX: number) => {
+      return zeroableArb.generator.flatMap((boundsY: number) => {
+        return Jsc.integer(boundsX, maxBounds).generator.flatMap((newX: number) => {
+          return Jsc.integer(boundsY, maxBounds).generator.flatMap((newY: number) => {
+            return zeroableArb.generator.flatMap((width: number) => {
+              return zeroableArb.generator.flatMap((height: number) => {
+                return Jsc.integer(newX + width, maxBounds).generator.flatMap((boundsW: number) => {
+                  return Jsc.integer(newY + height, maxBounds).generator.map((boundsH: number) => {
                     return {
                       newX,
                       newY,
@@ -45,8 +46,8 @@ UnitTest.test('BounderCalcRepositionTest', () => {
       const bounds = makeBounds(input.boundsX, input.boundsY, input.boundsW, input.boundsH);
       const output = Bounder.calcReposition(input.newX, input.newY, input.width, input.height, bounds);
 
-      const xIsVisible = output.limitX <= bounds.right() && output.limitX >= bounds.x();
-      const yIsVisible = output.limitY <= bounds.bottom() && output.limitY >= bounds.y();
+      const xIsVisible = output.limitX <= (bounds.right() - input.width) && output.limitX >= bounds.x();
+      const yIsVisible = output.limitY <= (bounds.bottom() - input.height) && output.limitY >= bounds.y();
       if (!xIsVisible) {
         return 'X is not inside bounds. Returned: ' + JSON.stringify(output);
       } else if (!yIsVisible) {