+++ modules/tinymce/src/core/main/ts/api/EventTypes.ts
@@ -21,7 +21,7 @@ export type SetContentEvent = SetContentArgs & { paste: boolean, selection: bool
 
 export type NewBlockEvent = { newBlock: Element };
 
-export type NodeChangedEvent = { element: Element, parents: Node[], selectionChange?: boolean };
+export type NodeChangedEvent = { element: Element, parents: Node[], selectionChange?: boolean, initial?: boolean };
 
 export type ObjectResizedEvent = { target: HTMLElement, width: number, height: number };