+++ modules/tinymce/src/plugins/searchreplace/main/ts/core/Actions.ts
@@ -7,6 +7,7 @@
 
 import { Element, Node } from '@ephox/dom-globals';
 import { Cell } from '@ephox/katamari';
+import { Pattern } from '@ephox/polaris';
 import DOMUtils from 'tinymce/core/api/dom/DOMUtils';
 import Editor from 'tinymce/core/api/Editor';
 import Tools from 'tinymce/core/api/util/Tools';
@@ -42,7 +43,7 @@ const markAllMatches = function (editor: Editor, currentSearchState: Cell<Search
 
   done(editor, currentSearchState, false);
 
-  return FindReplaceText.findAndReplaceDOMText(regex, node, marker, false, editor.schema);
+  return FindReplaceText.findAndReplaceDOMText(regex, node, marker, 1, editor.schema);
 };
 
 const unwrap = function (node: Node) {
@@ -122,7 +123,8 @@ const removeNode = function (dom: DOMUtils, node: Node) {
 
 const escapeSearchText = (text: string, wholeWord: boolean) => {
   const escapedText = text.replace(/[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, '\\$&').replace(/\s/g, '[^\\S\\r\\n]');
-  return wholeWord ? '\\b' + escapedText + '\\b' : escapedText;
+  const wordRegex = '(' + escapedText + ')';
+  return wholeWord ? `(?:^|\\s|${Pattern.punctuation()})` + wordRegex + `(?=$|\\s|${Pattern.punctuation()})` : wordRegex;
 };
 
 const find = function (editor: Editor, currentSearchState: Cell<SearchState>, text: string, matchCase: boolean, wholeWord: boolean) {